+++ cocos/scripting/js-bindings/auto/jsb_creator_camera_auto.cpp
@@ -146,7 +146,7 @@ extern se::Object* __jsb_cocos2d_Node_proto;
 
 static bool js_creator_CameraNode_finalize(se::State& s)
 {
-    CCLOG("jsbindings: finalizing JS object %p (creator::CameraNode)", s.nativeThisObject());
+    CCLOGINFO("jsbindings: finalizing JS object %p (creator::CameraNode)", s.nativeThisObject());
     creator::CameraNode* cobj = (creator::CameraNode*)s.nativeThisObject();
     if (cobj->getReferenceCount() == 1)
         cobj->autorelease();
