+++ cocos/scripting/js-bindings/manual/jsb_creator_manual.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace se {
+    class Object;
+}
+
+bool register_all_creator_manual(se::Object* obj);
+
